fix(front#1194) Доработка по событиям кпе вывод доп параметров

+++ src/app/widgets/EVJ/evj-events-workspace/components/evj-events-workspace-extra-options/evj-events-workspace-extra-options.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit, Output, EventEmitter, OnDestroy, OnChanges } from '@angular/core';
-import { IExtraOptionsWindow, IKpeAdditionalParameter } from "../../../../../dashboard/models/EVJ/events-widget";
+import { IExtraOptionsWindow, IKpeAdditionalParameter } from '../../../../../dashboard/models/EVJ/events-widget';
 import { EventsWorkspaceService } from '../../../../../dashboard/services/widgets/EVJ/events-workspace.service';
 import { AuthService } from '@core/service/auth.service';
 import { IAlertWindowModel } from '@shared/models/alert-window.model';
@@ -9,21 +9,21 @@ import {
     IKpeNotification,
     IKpeWorkspaceParameter,
 } from '../../../../../dashboard/models/EVJ/kpe-workspace.model';
-import { FormArray, FormBuilder, FormControl, FormGroup, Validators } from "@angular/forms";
+import { FormArray, FormBuilder, FormControl, FormGroup } from "@angular/forms";
 import { Subscription } from 'rxjs';
-
+import { IMessage } from '@shared/models/message.model';
 
 @Component({
-    selector: 'evj-evj-events-workspace-extra-options',
-    templateUrl: './evj-events-workspace-extra-options.component.html',
-    styleUrls: ['./evj-events-workspace-extra-options.component.scss'],
+    selector: "evj-evj-events-workspace-extra-options",
+    templateUrl: "./evj-events-workspace-extra-options.component.html",
+    styleUrls: ["./evj-events-workspace-extra-options.component.scss"]
 })
 export class EvjEventsWorkspaceExtraOptionsComponent implements OnInit, OnChanges, OnDestroy {
     @Input() public info: IExtraOptionsWindow = {
         data: null,
         isShow: false,
         acceptFunction: () => null,
-        closeFunction: () => null,
+        closeFunction: () => null
     };
     @Output() checked: any = new EventEmitter<boolean>();
 
@@ -34,17 +34,18 @@ export class EvjEventsWorkspaceExtraOptionsComponent implements OnInit, OnChange
     public notificationParametersData: IKpeNotification;
     public form: FormGroup;
     formArray: FormArray;
+    public index: number = -1;
+    public oldMessage: string = '';
 
     constructor(
         private ewService: EventsWorkspaceService,
         private kpeWorkspaceService: KpeWorkspaceService,
         private authService: AuthService,
         private formBuild: FormBuilder
-    ) {}
+    ) {
+    }
 
     ngOnInit(): void {
-        console.log('onInit');
-
         this.loadData();
         this.subscriptions.push(
             this.kpeWorkspaceService.showSelectParameters$.subscribe((res) => {
@@ -65,6 +66,8 @@ export class EvjEventsWorkspaceExtraOptionsComponent implements OnInit, OnChange
             this.loadData();
             this.kpeWorkspaceService.selectParameter$.next(this.info.data.selectedParameter);
             this.form.get('parameters').setValue(this.info.data?.selectedParameter);
+            this.formArray = this.form.get('dependentParameters') as FormArray;
+            this.formArray.clear();
             this.setDependentParameters(this.info.data?.dependentParameters);
         }
     }
@@ -76,12 +79,31 @@ export class EvjEventsWorkspaceExtraOptionsComponent implements OnInit, OnChange
     loadData(): void {
         if (this.ewService.event.id) {
             this.getParametersByNotification();
+            this.oldMessage = this.notificationDataToDescription(this.ewService.event.kpeAdditionalParameter);
+        } else {
+            this.oldMessage = this.notificationDataToDescription(this.notificationParametersData)
         }
         this.getParameters();
-        this.form = this.formBuild.group({
-            parameters: this.formBuild.control(null, Validators.required),
-            dependentParameters: this.formBuild.array([]),
-        });
+        this.getExtraParameters();
+        if (this.ewService.event.id && this.ewService.event.kpeAdditionalParameter) {
+            const controls = [];
+            this.ewService.event.kpeAdditionalParameter.dependentParameters.forEach((value) => {
+                controls.push(new FormGroup({
+                        dependentParameterId: new FormControl(value.dependentParameterId),
+                        numericValue: new FormControl(value.numericValue)
+                    }
+                ));
+            });
+            this.form = new FormGroup({
+                parameters: new FormControl(this.ewService.event.kpeAdditionalParameter.selectedParameter),
+                dependentParameters: new FormArray(controls)
+            });
+        } else {
+            this.form = new FormGroup({
+                parameters: new FormControl(null),
+                dependentParameters: new FormArray([])
+            });
+        }
     }
 
     get dependParameters(): FormArray {
@@ -96,12 +118,12 @@ export class EvjEventsWorkspaceExtraOptionsComponent implements OnInit, OnChange
     }
 
     createFormGroup(
-        dependentParameterId: number = this.dependentParameters[0]?.id,
+        dependentParameterId: number = this.dependentParameters[0].id,
         numericValue: number = 0
     ): FormGroup {
         return this.formBuild.group({
             dependentParameterId,
-            numericValue,
+            numericValue
         });
     }
 
@@ -136,18 +158,26 @@ export class EvjEventsWorkspaceExtraOptionsComponent implements OnInit, OnChange
 
     private async deleteParametersByNotification(): Promise<void> {
         try {
-            await this.kpeWorkspaceService.deleteKpeNotificationParameters(this.ewService.event);
-            this.ewService.event.kpeAdditionalParameter = null;
-
+            if (this.ewService.event.id) {
+                await this.kpeWorkspaceService.deleteKpeNotificationParameters(this.ewService.event);
+                this.ewService.event.kpeAdditionalParameter = null;
+                this.ewService.event.facts.splice(this.index, 1);
+                this.index = -1;
+            } else {
+                this.ewService.event.kpeAdditionalParameter = null;
+                this.ewService.event.facts.splice(this.index, 1);
+                this.index = -1;
+                this.ewService.acceptButton$.next(null);
+            }
             const popupWindow: IExtraOptionsWindow = {
                 isShow: false,
-                type: 'reset',
+                type: 'reset'
             };
             this.ewService.extraOptionsWindow$.next(popupWindow);
         } catch {
             const popupWindow: IExtraOptionsWindow = {
                 isShow: false,
-                type: 'save',
+                type: 'save'
             };
             this.ewService.extraOptionsWindow$.next(popupWindow);
         }
@@ -157,7 +187,7 @@ export class EvjEventsWorkspaceExtraOptionsComponent implements OnInit, OnChange
     public cancel(): void {
         const popupWindow: IExtraOptionsWindow = {
             isShow: false,
-            type: 'cancel',
+            type: 'cancel'
         };
         this.ewService.extraOptionsWindow$.next(popupWindow);
         this.formArray?.clear();
@@ -183,44 +213,44 @@ export class EvjEventsWorkspaceExtraOptionsComponent implements OnInit, OnChange
             return {
                 numericValue: value.numericValue,
                 dependentParameterId: id,
+                name: this.dependentParameters[value.dependentParameterId - 1].name
             };
         });
         this.notificationParametersData = {
             selectedParameter: this.form.value.parameters,
             dependentParameters,
             createdAt: new Date(),
-            createdBy: this.authService.user$.getValue().id,
+            createdBy: this.authService.user$.getValue().id
         };
+        this.ewService.event.facts.forEach((v, index) => {
+            if (v.comment === this.oldMessage) {
+                this.index = index;
+                console.log(index);
+                this.oldMessage = v.comment;
+            }
+        });
         try {
             if (this.ewService.event.id) {
                 const res = await this.kpeWorkspaceService.postKpeNotificationParameters(
                     this.ewService.event,
                     this.notificationParametersData
                 );
-                this.ewService.event.kpeAdditionalParameter = {
-                    selectedParameterId: res.selectedParameter.id,
-                    dependentParameters: res.dependentParameters,
-                    createdAt: res.createdAt,
-                    createdBy: res.createdBy
-                };
+                this.ewService.event.kpeAdditionalParameter = res;
+                this.addKpeAdditionalToChat(res);
             } else {
-                this.ewService.event.kpeAdditionalParameter = {
-                    selectedParameterId: this.notificationParametersData.selectedParameter.id,
-                    dependentParameters: this.notificationParametersData.dependentParameters,
-                    createdAt: this.notificationParametersData.createdAt,
-                    createdBy: this.notificationParametersData.createdBy
-                }
-                this.ewService.acceptButton$.next(this.notificationParametersData)
+                this.ewService.event.kpeAdditionalParameter = this.notificationParametersData;
+                this.addKpeAdditionalToChat(this.notificationParametersData);
+                console.log(this.notificationParametersData);
+                this.ewService.acceptButton$.next(this.notificationParametersData);
             }
         } catch (error) {
             console.error(error);
         }
         const popupWindow: IExtraOptionsWindow = {
             type: 'save',
-            isShow: false,
+            isShow: false
         };
         this.ewService.extraOptionsWindow$.next(popupWindow);
-        console.log(this.ewService.event.kpeAdditionalParameter);
     }
 
     //  DELETE Удалить данные
@@ -231,18 +261,43 @@ export class EvjEventsWorkspaceExtraOptionsComponent implements OnInit, OnChange
             acceptText: 'Да',
             cancelText: 'Нет',
             acceptFunction: () => this.deleteParametersByNotification(),
-            closeFunction: () => this.ewService.ewAlertInfo$.next(null),
+            closeFunction: () => this.ewService.ewAlertInfo$.next(null)
         };
         this.ewService.ewAlertInfo$.next(alertWindow);
         const popupWindow: IExtraOptionsWindow = {
             isShow: false,
-            type: 'reset',
+            type: 'reset'
         };
         this.ewService.extraOptionsWindow$.next(popupWindow);
     }
 
+    public addKpeAdditionalToChat(data: IKpeAdditionalParameter): void {
+        console.log(this.index);
+        if (this.index > -1) {
+            this.ewService.event.facts[this.index].comment = this.notificationDataToDescription(data);
+        } else {
+            this.ewService.event.facts.push({
+                comment: this.notificationDataToDescription(data)
+            } as IMessage);
+        }
+    }
+
+    public notificationDataToDescription(data: IKpeAdditionalParameter): string {
+        const description: string[] = [];
+        let result: string = '';
+
+        description.push(data?.selectedParameter?.name);
+        data?.dependentParameters?.forEach((value) => {
+            value.unit ?
+                description.push(value.name + ' ' + value?.numericValue?.toString() + ' ' + value?.unit?.toString())
+                : description.push(value.name + ' ' + value?.numericValue?.toString());
+        });
+        result = '\n'.concat(description.join(';\n'));
+        return result;
+    }
+
     public compareFnDependent(a, b): boolean {
-        return a?.dependentParameterId === b?.dependentParameterId;
+        return a === b;
     }
 
     public compareFn(a, b): boolean {
